// src/types/cup-color.ts
export type CupColor =
  | "green"
  | "white"
  | "orange"
  | "red"
  | "yellow"
  | "purple"
  | "blue"
  | "pink";

export const colorMap = {
  blue: "bg-blue-500",
  red: "bg-red-500",
  green: "bg-green-500",
  yellow: "bg-yellow-500",
  purple: "bg-purple-500",
  white: "bg-white",
  pink: "bg-pink-500",
  orange: "bg-orange-500",
};